client/beefy: decouple voter init from aux db state load
